tidy up the recoil things a little
