+++ src/components/cores/info/notInstalled.tsx
@@ -2,7 +2,7 @@ import { useMemo } from "react"
 import { useRecoilValue } from "recoil"
 import { useInstallCore } from "../../../hooks/useInstallCore"
 import { useInventoryItem } from "../../../hooks/useInventoryItem"
-import { DownloadURLSelectorFamily } from "../../../recoil/selectors"
+import { DownloadURLSelectorFamily } from "../../../recoil/inventory/selectors"
 import { Controls } from "../../controls"
 import { Link } from "../../link"
 import { InstallOptions } from "./installOptions"
