+++ src/recoil/github/selectors.ts
@@ -1,0 +1,18 @@
+import { GithubRelease } from "../../types"
+import { selectorFamily } from "recoil"
+
+export const GithubReleasesSelectorFamily = selectorFamily<
+  GithubRelease[],
+  { owner: string; repo: string }
+>({
+  key: "GithubReleasesSelectorFamily",
+  get:
+    ({ owner, repo }) =>
+    async () => {
+      const response = await fetch(
+        `https://api.github.com/repos/${owner}/${repo}/releases`
+      )
+
+      return (await response.json()) as GithubRelease[]
+    },
+})
